+++ maptasker/src/getputer.py
@@ -198,7 +198,9 @@ def process_old_formatted_file(
             temp_args = program_arguments["backup_file_http"].split(":")
             program_arguments["android_ipaddr"] = temp_args[1][2:]
             program_arguments["android_port"] = temp_args[2]
-            program_arguments["android_file"] = program_arguments["backup_file_location"]
+            program_arguments["android_file"] = program_arguments[
+                "backup_file_location"
+            ]
             del program_arguments["backup_file_http"]
             del program_arguments["backup_file_location"]
 
@@ -245,7 +247,9 @@ def read_toml_file(new_file: str) -> tuple[dict, dict]:
 
             # Program arguments
             try:
-                program_arguments = settings["program_arguments"]  # Get the program arguments
+                program_arguments = settings[
+                    "program_arguments"
+                ]  # Get the program arguments
             except KeyError:
                 program_arguments = initialize_runtime_arguments()
             try:
@@ -319,8 +323,10 @@ def save_restore_args(
         :param to_save: True if this is a save request, False is restore request
         :return: program runtime arguments saved/restored, colors to use saved/restored
     """
-    new_file = f"{Path.cwd()}{PrimeItems.slash}{ARGUMENTS_FILE}"
-    old_file = f"{Path.cwd()}{PrimeItems.slash}{OLD_ARGUMENTS_FILE}"
+    our_path = Path.cwd()
+    print("bingo", our_path)
+    new_file = f"{our_path}{PrimeItems.slash}{ARGUMENTS_FILE}"
+    old_file = f"{our_path}{PrimeItems.slash}{OLD_ARGUMENTS_FILE}"
 
     # Saving?
     if to_save: