+++ maptasker/src/sysconst.py
@@ -320,8 +320,9 @@ GEMINI_MODELS = [
     "gemini-1.5-pro",
     "gemini-2.0-flash",
     "gemini-2.0-flash-lite",
-    "gemini-2.5-flash-preview-05-20",
-    "gemini-2.5-pro-preview-05-06",
+    "gemini-2.5-flash",
+    "gemini-2.5-flash-lite-preview-06-17",
+    "gemini-2.5-pro",
 ]
 MODEL_GROUPS = {
     "OpenAI": OPENAI_MODELS,